pyproject.toml
[tool.poetry]
name = "whatsapp-invoice-assistant"
version = "0.1.0"
description = "AI-powered WhatsApp invoice assistant for processing and managing invoice data"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.104.0"
uvicorn = "^0.23.2"
sqlalchemy = "^2.0.22"
alembic = "^1.12.0"
pydantic = "^2.4.2"
pydantic-settings = "^2.0.3"
celery = "^5.3.4"
redis = "^5.0.1"
twilio = "^8.9.1"
boto3 = "^1.28.62"
openai = "^1.1.1"
python-multipart = "^0.0.6"
jinja2 = "^3.1.2"
httpx = "^0.25.0"
pillow = "^10.0.1"
python-dotenv = "^1.0.0"
asyncpg = "^0.28.0"
psycopg2-binary = "^2.9.9"
langchain = "^0.1.0"
langgraph = "^0.0.25"
pyyaml = "^6.0.1"
fpdf = "^1.7.2"
pgvector = "^0.2.1"
sentence-transformers = "^2.2.2"
numpy = "^1.26.0"
langchain-core = "^0.1.53"
pymongo = "^4.6.1"
motor = "^3.3.2"
langsmith = "^0.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
pytest-asyncio = "^0.21.1"
black = "^23.9.1"
flake8 = "^6.1.0"
mypy = "^1.5.1"
pre-commit = "^3.5.0"
pytest-cov = "^4.1.0"
faker = "^19.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py39"]
include = '\.pyi?$'

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
strict_optional = true

[[tool.mypy.overrides]]
module = ["twilio.*", "celery.*", "boto3.*", "langchain.*", "langgraph.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_scope = "function"
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
log_cli = true
log_cli_level = "DEBUG" 